refactor: move connections into their own components folder
