+++ examples/fip_behavior_instrument.py
@@ -29,7 +29,8 @@ from aind_data_schema.components.devices import (
     Tube,
     Computer,
 )
-from aind_data_schema.core.instrument import Instrument, Connection, ConnectionData, ConnectionDirection
+from aind_data_schema.components.connections import Connection, ConnectionData, ConnectionDirection
+from aind_data_schema.core.instrument import Instrument
 from aind_data_schema.components.identifiers import Software
 from aind_data_schema.components.coordinates import (
     CoordinateSystemLibrary,
