+++ examples/multiplane_ophys_instrument.py
@@ -33,7 +33,8 @@ from aind_data_schema.components.devices import (
 )
 from aind_data_schema.components.devices import Objective
 from aind_data_schema.components.identifiers import Software
-from aind_data_schema.core.instrument import Instrument, Connection, ConnectionData, ConnectionDirection
+from aind_data_schema.components.connections import Connection, ConnectionData, ConnectionDirection
+from aind_data_schema.core.instrument import Instrument
 
 instrument = Instrument(
     location="429",
