+++ tests/test_acquisition.py
@@ -21,7 +21,7 @@ from aind_data_schema.components.configs import (
 from aind_data_schema.components.coordinates import Affine, CoordinateSystemLibrary, Scale, Translation
 from aind_data_schema.components.identifiers import Person
 from aind_data_schema.core.acquisition import Acquisition, AcquisitionSubjectDetails, DataStream
-from aind_data_schema.core.instrument import Connection
+from aind_data_schema.components.connections import Connection
 from examples.ephys_acquisition import acquisition as ephys_acquisition
 from examples.exaspim_acquisition import acq as exaspim_acquisition
