+++ tests/test_metadata.py
@@ -17,7 +17,8 @@ from aind_data_schema.components.subjects import BreedingInfo, Housing, MouseSub
 from aind_data_schema.components.surgery_procedures import BrainInjection
 from aind_data_schema.core.acquisition import Acquisition, AcquisitionSubjectDetails, DataStream
 from aind_data_schema.core.data_description import DataDescription, Funding
-from aind_data_schema.core.instrument import Connection, Instrument
+from aind_data_schema.components.connections import Connection
+from aind_data_schema.core.instrument import Instrument
 from aind_data_schema.core.metadata import Metadata, create_metadata_json
 from aind_data_schema.core.procedures import Procedures, Surgery
 from aind_data_schema.core.processing import DataProcess, Processing, ProcessName, ProcessStage
